Extract Carousel Logic into Hook
Need to pass in number of items in carousel and apply the xTranslation on the parent box containing the motion container and items, apply the itemRef on a item in the carousel to determine the width of each item, and optionally apply the handleHoverDuration to change the speed of the carousel on hover.
